404ERROR/requirements.txt
pandas
matplotlib
flask
flask_login
flask_migrate
flask_sqlalchemy
mysql-connector-python
requests
beautifulsoup4
pymysql
chardet